Add files via upload
Re-named some functions, and added in functionality that allows the user to choose to Save the file or not. If they don't Save, a message will appear, and the program exits. If they do Save, it will save the CSV. Allowing the user to decide the file location to save is coming next.